Complete 'delaunay' function
