+++ requirements.txt
@@ -1,5 +1,6 @@
 bentley_ottmann>=0.8.0
+decision>=0.1.0
 dendroid>=0.4.2
 hypothesis>=4.55.0
 reprit>=0.2.2
-robust>=0.2.1
+robust>=0.2.4
